refactor: tests, fix parser error exception
