+++ src/fdk_rdf_parser/classes/exceptions.py
@@ -2,7 +2,12 @@ from rdflib.exceptions import ParserError as RdflibParserError
 
 
 class ParserError(RdflibParserError):
-    pass
+    def __init__(self, msg: str = "Parser error"):
+        RdflibParserError.__init__(self, msg)
+        self.msg: str = msg
+
+    def __str__(self) -> str:
+        return self.msg
 
 
 class MultipleResourcesError(ValueError):
